api/src/requirements/base.txt
aiomysql==0.2.0
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
fastapi==0.110.2
h11==0.14.0
idna==3.7
iso8601==1.1.0
pydantic==2.7.0
pydantic_core==2.18.1
pypika-tortoise==0.1.6
python-dotenv==1.0.1
pytz==2024.1
sniffio==1.3.1
starlette==0.37.2
tortoise-orm==0.20.0
typing_extensions==4.11.0
uvicorn==0.29.0
